the driver code is ready for testing,
test include:
	-driver should get the rom data from the DRT by itself

need to write a simple tty platform device that will interact
with the sycamore bus.

+++ xylem/driver/sycamore/sycamore_control.h
@@ -18,7 +18,7 @@
 
 int sycamore_write(sycamore_t * sycamore, 
 					u32 command,
-					u32 device_address,
+					u8 device_address,
 					u32 offset,
 					char *buffer, u32 length);
